backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
opencv-python==4.8.0.76
pandas==2.0.3
matplotlib==3.7.2
fpdf==1.7.2
joblib==1.3.2
shap==0.42.3
aiofiles==23.2.1
python-multipart==0.0.6
cors==1.0.1
ultralytics==8.0.234
pillow==10.0.0
